+++ src/inputs/MultiSelectField.tsx
@@ -1,10 +1,9 @@
 import { ReactNode } from "react";
 import { Value } from "src/inputs";
-import { BeamSelectFieldBaseProps, SelectFieldBase } from "src/inputs/internal/SelectFieldBase";
+import { ComboBoxBase, ComboBoxBaseProps } from "src/inputs/internal/ComboBoxBase";
 import { HasIdAndName, Optional } from "src/types";
 
-export interface MultiSelectFieldProps<O, V extends Value>
-  extends Exclude<BeamSelectFieldBaseProps<O, V>, "unsetLabel"> {
+export interface MultiSelectFieldProps<O, V extends Value> extends Exclude<ComboBoxBaseProps<O, V>, "unsetLabel"> {
   /** Renders `opt` in the dropdown menu, defaults to the `getOptionLabel` prop. */
   getOptionMenuLabel?: (opt: O) => string | ReactNode;
   getOptionValue: (opt: O) => V;
@@ -37,7 +36,7 @@ export function MultiSelectField<O, V extends Value>(
   } = props;
 
   return (
-    <SelectFieldBase
+    <ComboBoxBase
       multiselect
       {...otherProps}
       options={options}
